pr comments resolved. 2 issues still require discussion

+++ BeerDispancer/BeerDispencer.Application/DTO/DispencerDto.cs
@@ -6,7 +6,7 @@ namespace BeerDispancer.Application.DTO
 	public class DispencerDto
 	{
         public Guid Id { get; set; }
-        public double? Volume { get; set; }
+        public decimal? Volume { get; set; }
         public DispencerStatus? Status { get; set; }
     }
 }
