+++ BeerDispancer/BeerDispencer.Domain/Implementations/BeerFlowSettings.cs
@@ -4,8 +4,8 @@ namespace BeerDispencer.Domain.Implementations
 {
     public class BeerFlowSettings : IBeerFlowSettings
     {
-        public double LitersPerSecond { get; set; }
-        public double PricePerLiter { get; set; }
+        public decimal LitersPerSecond { get; set; }
+        public decimal PricePerLiter { get; set; }
 
     }
 }
