+++ BeerDispancer/BeerDispencer.Domain/Implementations/Calculator.cs
@@ -11,7 +11,7 @@ namespace BeerDispencer.Domain.Implementations
             _settings = settings;
         }
 
-        public  double? GetFlowVolume(DateTime? closedAt, DateTime? openAt)
+        public  decimal? GetFlowVolume(DateTime? closedAt, DateTime? openAt)
         {
             var duration = closedAt - openAt;
 
@@ -20,10 +20,11 @@ namespace BeerDispencer.Domain.Implementations
                 return null;
             }
 
-            return duration.Value.TotalSeconds * _settings.LitersPerSecond;
+            return (decimal)duration.Value.TotalSeconds * _settings.LitersPerSecond;
         }
 
-        public  double? GetTotalSpent(double? volume)
+
+        public  decimal? GetTotalSpent(decimal? volume)
         {
             return volume * _settings.PricePerLiter;
         }
