+++ BeerDispancer/BeerDispencer.Infrastructure/Persistence/Entities/Dispencer.cs
@@ -10,7 +10,7 @@ namespace BeerDispencer.Infrastructure.Persistence.Entities
     {
         [Key]
         public Guid Id { get; set; }
-        public double? Volume { get; set; }
+        public decimal? Volume { get; set; }
         public DispencerStatus? Status { get; set; }
     }
 }
